ai-backend-flask/requirements.txt
# Core Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0

# AI and OpenAI
openai>=1.3.0
langchain>=0.0.300

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
chardet>=5.0.0

# Database
# sqlite3 is built into Python

# Utilities
requests>=2.31.0
python-multipart>=0.0.6
Werkzeug>=2.3.0

# Logging
loguru>=0.7.0

# Development
python-dotenv>=1.0.0 